requirements.txt
MetaTrader5
pyzmq
pandas
numpy
streamlit
plotly
pandas_ta
scipy
streamlit-lightweight-charts
xgboost
matplotlib
scikit-learn
requests